Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: Ic11a02b94cb9708a8b830942264e5ee80b748a5e
(cherry picked from commit e27bc2c072e26c2c4659a57bcefc16506fc7ed69)